refactor(warframe): 优化入侵信息接口并添加缓存支持
- 修改 Invasions 类中的 completion 字段类型从 Int 改为 Double
- 在 ParseDataUtil 中添加对入侵信息的缓存支持
- 优化 WarframeController 中的 invasions 接口，使用缓存数据
- 统一使用 GET 请求替换原有的 RRequestMapping 注解

+++ wuliang-warframe/src/main/kotlin/bot/wuliang/moudles/Invasions.kt
@@ -53,7 +53,7 @@ data class Invasions(
     /**
      * 入侵的完成百分比。如果达到 0，防守方获胜 Grineer vs. Corpus 入侵从 50 开始，Infested 入侵从 100 开始
      */
-    var completion: Int? = null,
+    var completion: Double? = null,
 
     /**
      * 入侵是否结束
